Superhirn PATCH by Florian Zeller: Turned some implicit conversions into explicit convertions
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@10077 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/SUPERHIRN/ClusteredMS2ConsensusSpectrum.h
@@ -134,7 +134,7 @@ public:
   // function to access the
   // variables of the class
  
-  int getNumberOfSpectraScan(){return MS2Scans.size();};
+  int getNumberOfSpectraScan(){return (int) MS2Scans.size();};
   std::vector<int>::iterator getSpectraScanNumberStart(){return MS2Scans.begin();};
   std::vector<int>::iterator getSpectraScanNumberEnd(){return MS2Scans.end();};
